packages/create-aethermind-app/templates/python/requirements.txt
aethermind-sdk>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
